requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# OpenAI for optional LLM features
openai==1.10.0

# Additional utilities
python-dotenv==1.0.0
